Add files via upload
These are some javascript projects that I wrote when I was in highschool. 
"Interceptor" is a game where one uses the up&down, and left&right arrows on the keyboard to steer a black 'Interceptor' that is supposed to catch the "falling" blue squares. The game is lost when a blue square touches the bottom of the playing field. 
"FLOPPY_BIRD" is a game where one uses the up key on the keyboard to control a small square (a "bird") in order to fly through obstacles without touching them. 
"Conway's game of life" is not a game, but rather a cellular automaton developed by the british mathematician John Horton Conway (1937 - 2020). There are specific rules to how it operates ( https://en.wikipedia.org/wiki/Conway's_Game_of_Life ), however, the way I implemented it does not account for boundary conditions at the borders of the playing field.